增加回复功能

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -1,8 +1,8 @@
 package life.majiang.community.controller;
 
 import life.majiang.community.model.Question;
-import life.majiang.community.model.User;
 import life.majiang.community.dto.QuestionDTO;
+import life.majiang.community.model.User;
 import life.majiang.community.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -20,7 +20,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,Model model){
+    public String edit(@PathVariable(name = "id") Long id,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
         model.addAttribute("description",question.getDescription());
@@ -39,7 +39,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam(value="id",required = false) Integer id,
+            @RequestParam(value="id",required = false) Long id,
             HttpServletRequest request,
             Model model
     ){
